phase-5/services/chat-api/requirements.txt
# Chat API Service Dependencies - Phase V
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlmodel==0.0.14
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# Dapr HTTP client
httpx==0.25.2

# Data validation
pydantic==2.5.2

# Date/time handling for recurrence
python-dateutil==2.8.2

# Structured logging
structlog==23.3.0

# Environment variables
python-dotenv==1.0.0

# Testing (for future use)
pytest==7.4.3
pytest-asyncio==0.21.1

# Code quality
ruff==0.1.8
black==23.12.1
